TimeFrame class constructor and match method made inline; fixed Windows compiler warnings; some cosmetic changes

+++ src/server/core/netmap_link.cpp
@@ -190,7 +190,7 @@ bool NetworkMapLink::update(const ObjLink& src)
          {
             while((startIndex > 0) && _istspace(config.charAt(startIndex - 1)))
                startIndex--;
-            int endIndex = config.find(_T("</objectStatusList>"));
+            ssize_t endIndex = config.find(_T("</objectStatusList>"));
             if (endIndex != -1)
             {
                endIndex += 19;
